+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- Template Viewer: a new component for managing templates to just have a view of the template without editing it.
+
+### Changes
+
+- Manage Templates: Users can now create transactions from templates even if they are not in quick view.
+
 ## [0.6.0]
 
 ### Added